Removed duplicated admin helpers
no ref

Five small duplications and oddities in apps/admin, collapsed with no
behavior change:

- Extracted the optimistic newsletter reorder logic duplicated between
  newsletters.tsx and newsletters/newsletters-tab-content.tsx into a
  shared use-newsletter-reorder hook used by both.
- Extracted the hand-rolled x-ghost-preview POST duplicated between
  theme-preview.tsx and announcement-bar-preview.tsx into a shared
  fetchFrontendPreview helper. It targets the site front-end, not the
  Admin API, so it stays a plain fetch.
- Replaced the useQuery in useWhatsNew, whose queryFn was a synchronous
  boolean over already-loaded data, with a useMemo over the same inputs,
  and updated its consumers and tests to the plain return shape.
- Replaced the hand-rolled email regexes in member-detail-edit.ts and
  import-members/mapping.ts with validator.isEmail, which the rest of
  Admin and the server-side member validation already use. No test
  fixture depended on the looser hand-rolled match.
- Collapsed the four @xyflow/react/dist/style.css imports in the
  automations canvas to automation-canvas.tsx, the one module that
  mounts ReactFlow.
