+++ apps/admin/src/automations/components/canvas/controls.tsx
@@ -1,4 +1,3 @@
-import '@xyflow/react/dist/style.css';
 import React, { useState } from 'react';
 import {
   Button,
