+++ apps/admin/src/layout/app-sidebar/user-menu.tsx
@@ -118,7 +118,7 @@ interface UserMenuProps extends React.ComponentProps<typeof DropdownMenu> {
 }
 function UserMenu(props: UserMenuProps) {
   const currentUser = useCurrentUser();
-  const { data: whatsNewData } = useWhatsNew();
+  const { hasNew } = useWhatsNew();
   const { showUpgradeBanner } = useUpgradeStatus();
 
   return (
@@ -131,7 +131,7 @@ function UserMenu(props: UserMenuProps) {
         >
           <div className="relative">
             <UserMenuAvatar />
-            {whatsNewData?.hasNew && (
+            {hasNew && (
               <span className="absolute -top-0.5 -right-0.5">
                 <Indicator
                   data-testid="whats-new-avatar-badge"
@@ -172,7 +172,7 @@ function UserMenu(props: UserMenuProps) {
         >
           <LucideIcon.Sparkles />
           <UserMenuItem.Label>What’s new?</UserMenuItem.Label>
-          {whatsNewData?.hasNew && (
+          {hasNew && (
             <div className="flex flex-1 justify-end">
               <Indicator
                 data-testid="whats-new-menu-badge"
