+++ apps/admin/src/settings/utils/fetch-frontend-preview.ts
@@ -1,0 +1,18 @@
+// Fetches a site front-end page rendered with the given `x-ghost-preview` data.
+// This targets the front-end, not the Admin API, so it stays a plain fetch.
+export function fetchFrontendPreview(url: string, previewData: string): Promise<string> {
+  // Suppress the admin toolbar in previews
+  const previewUrl = new URL(url);
+  previewUrl.searchParams.set('admin_toolbar', '0');
+
+  return fetch(previewUrl.toString(), {
+    method: 'POST',
+    headers: {
+      'Content-Type': 'text/html;charset=utf-8',
+      'x-ghost-preview': previewData,
+      Accept: 'text/html',
+    },
+    mode: 'cors',
+    credentials: 'include',
+  }).then((response) => response.text());
+}
